Setup tools and package setup
- Added setuptools
- Split into api and cli packages for better maintainance and
  extendability
- Split util module into modules for each api action

Closes #5
