+++ setup.cfg
@@ -1,0 +1,13 @@
+[metadata]
+name = shinyhunter
+version = 0.1.0
+
+[options]
+packages = find:
+package_dir =
+    =src
+
+[options.entry_points]
+console_scripts =
+    shinyhunter = cli.cli:cli
+
